Add files via upload
This week's homework was fairly easy, as long as you are good at wrapping your head around the coordinates. For me, it's difficult to remember that if you want a line directly within the rectangle, you'd actually have to add half the distance of the coordinate on to the starting point, because the rectangle is plotted using the left top point, not its center.